set limit to table desc and changing amount to float type with step size of 5.0

+++ php/setEntry.php
@@ -13,7 +13,7 @@
                                      edate date not null, 
                                      category int, 
                                      description varchar(255), 
-                                     amount int, 
+                                     amount float, 
                 Primary key(idx))";
     } else {
         $sql = "TRUNCATE TABLE $table";
